requirements.txt
Flask==3.1.0
flask-cors==5.0.1
pymongo==4.11.3
python-dotenv==1.0.1
PyJWT==2.10.1
reportlab==4.2.5
Werkzeug==3.1.3
